qa: add args.out_url
This option will allow users to post report data to any pull request.

We currently have an argument called -ogh. This also allows users to
post report data to github. However, the pull request or github dir
must contain font binaries.

This arg is more flexible since fonts are not required. It allows users
to build fonts using a CI such as travis and then run gftools on the
built fonts and post the results back to the pr.