Add sample_glyphs and sample_text fields to FamilyProto and remove fallback_language field from SampleTextProto.
Sample glyphs are a per-family field since not all glyphs associated with a given language will be included in a given font.

Sample text can be specified at the family level as an override to any language sample text or in cases where no language is supported, e.g. bar code fonts, symbols, music, etc.

Fallback language does not need to be specified in the per-language SampleTextProto anymore. Our new strategy is to cycle through the available sample texts of other languages of the same script. This eliminates the need for manually specified fallback languages.