Make output more deterministic
Currently, the output of the merge process is non-deterministic: if the
process is run twice with the same input UFOs, the output will differ.

This is because the existing order of properties is discarded in the
merge process when values are represented as sets. While this yields
semantically identical results, it also creates lengthy and unnecessary
diffs, which evade normalisation.

This commit converts the sets back into lists through `sorted()` instead
of `list()`, to give the same output across multiple runs for the same
input.