[builder] Fix building static instances when glyphData is used
The glyphData option is only for Glyphs sources, but when building
statics, we create UFO instances from Glyphs sources first, then build
the UFOs, so --glyph-data fontmake option ends up being applied to UFO
sources and it fails.

We now skip glyphData when building statics, and pass it to the
instantiateUfo operation instead.