[fontc] Only crater needs unique build file names
When running in the context of crater, gftools builder may be invoked
multiple times in parallel, which can cause the generated ninja.build
files to overwrite each other.

To fix this, I implemented a scheme where we generate a unique name for
the generated build file, so that multiple could exist and be in use for
the same project, at the same time.

These files can pile up, though; I had initially thought they were
cleaned up by default, but that doesn't seem to be the case.

This patch makes it so we only use this new naming scheme if we're
running for crater, and uses the old names otherwise.

It also makes it so we always set the cleanup flag if we're running for
crater.