Raise an error when an unknown fontmake filter is provided
My understanding is that this was the original intention of the code,
but that the fallthrough case performed the task of skipping too.

Merging would lead to additional fails on crater, but in at least some
cases (e.g. the Jersey family) for families that have a low % match
already owing to an unsupported filter; this would become more explicit.