gftools.builder: allow to pass --experimental-extra-arg down to fontc and fontmake
this is useful for ttx_diff.py to control e.g. whether to pass the --no-production-names flag to fontc and fontmake, we want to rename glyphs to their final production names by default when comparing fontc vs fontmake output but sometimes it useful to keep original human-readable glyph names for debugging.